# packages/mail-api/openapi/mail-proxy.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mail Proxy API",
    "version": "0.1.0",
    "description": "REST facade over Mailcow IMAP/SMTP and SOGo CalDAV"
  },
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session token"
      }
    },
    "schemas": {
...
      "MailSessionRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "MailSessionResponse": {
        "type": "object",
        "required": ["sessionToken", "expiresAt", "email"],
        "properties": {
          "sessionToken": { "type": "string" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "email": { "type": "string", "format": "email" }
        }
      },
...
      "MailFoldersResponse": {
        "type": "object",
        "required": ["folders"],
        "properties": {
          "folders": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MailFolder" }
          },
          "delimiter": { "type": "string" }
        }
      },
      "CreateMailFolderRequest": {
        "type": "object",
        "required": ["path"],
        "properties": {
          "path": { "type": "string" }
        }
      },
...
      "MailSessionExchangeRequest": {
        "type": "object",
        "required": ["email", "realmUrl", "apiKey"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "realmUrl": { "type": "string", "format": "uri" },
          "apiKey": { "type": "string" },
          "password": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid session",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/mail/session/exchange": {
      "post": {
        "operationId": "exchangeMailSession",
        "summary": "Exchange Zulip credentials for mail session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MailSessionExchangeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MailSessionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/v1/mail/session": {
      "post": {
        "operationId": "createMailSession",
        "summary": "Create mail session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MailSessionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MailSessionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      },
      "delete": {
        "operationId": "deleteMailSession",
        "summary": "Delete mail session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Session deleted" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/folders": {
      "get": {
        "operationId": "listMailFolders",
        "summary": "List mail folders",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Folder list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MailFoldersResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "createMailFolder",
        "summary": "Create mail folder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMailFolderRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Folder created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PathResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "patch": {
        "operationId": "renameMailFolder",
        "summary": "Rename mail folder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RenameMailFolderRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder renamed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PathResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "operationId": "deleteMailFolder",
        "summary": "Delete mail folder",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "path", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "delimiter", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FolderPathRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Folder deleted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/folders/move": {
      "post": {
        "operationId": "moveMailFolder",
        "summary": "Move mail folder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MoveMailFolderRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder moved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PathResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/folders/clear": {
      "post": {
        "operationId": "clearMailFolder",
        "summary": "Clear all messages in folder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FolderPathRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder cleared",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OkResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/folders/mark-all-read": {
      "post": {
        "operationId": "markAllMailFolderRead",
        "summary": "Mark all messages in folder as read",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FolderPathRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder marked read",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OkResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/messages": {
      "get": {
        "operationId": "listMailMessages",
        "summary": "List messages in folder",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "folder", "in": "query", "schema": { "type": "string", "default": "INBOX" } },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Message list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MailMessagesResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "sendMailMessage",
        "summary": "Send mail message",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendMailRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OkResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/messages/{uid}": {
      "get": {
        "operationId": "getMailMessage",
        "summary": "Get message detail",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          },
          { "name": "folder", "in": "query", "schema": { "type": "string", "default": "INBOX" } },
          { "name": "markSeen", "in": "query", "schema": { "type": "boolean", "default": true } }
        ],
        "responses": {
          "200": {
            "description": "Message detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MailMessageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "patchMailMessageFlags",
        "summary": "Update message flags",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MailMessageFlagsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flags updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OkResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "operationId": "deleteMailMessage",
        "summary": "Delete message",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          },
          { "name": "folder", "in": "query", "schema": { "type": "string", "default": "INBOX" } }
        ],
        "responses": {
          "204": { "description": "Message deleted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/messages/{uid}/attachments": {
      "get": {
        "operationId": "listMailMessageAttachments",
        "summary": "List message attachments",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          },
          { "name": "folder", "in": "query", "schema": { "type": "string", "default": "INBOX" } }
        ],
        "responses": {
          "200": {
            "description": "Attachment list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MailAttachmentsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/mail/messages/{uid}/attachments/{attachmentId}": {
      "get": {
        "operationId": "getMailMessageAttachment",
        "summary": "Download message attachment",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "attachmentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          { "name": "folder", "in": "query", "schema": { "type": "string", "default": "INBOX" } }
        ],
        "responses": {
          "200": {
            "description": "Attachment binary",
            "content": {
              "application/octet-stream": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/mail/search": {
      "get": {
        "operationId": "searchMailMessages",
        "summary": "Search messages",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "folder", "in": "query", "schema": { "type": "string" } },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MailSearchResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/sync": {
      "get": {
        "operationId": "syncMailFolder",
        "summary": "Sync folder changes since UID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "folder", "in": "query", "schema": { "type": "string", "default": "INBOX" } },
          { "name": "sinceUid", "in": "query", "schema": { "type": "integer", "minimum": 0 } }
        ],
        "responses": {
          "200": {
            "description": "Sync delta",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MailSyncResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/messages/batch": {
      "post": {
        "operationId": "batchMailMessages",
        "summary": "Batch message operations",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MailBatchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch completed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OkResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/mail/drafts": {
      "post": {
        "operationId": "createMailDraft",
        "summary": "Create mail draft",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendMailRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MailMessageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}